refactor(product): consolidate product validation logic and remove validator package
